fixing base stats etc

+++ src/main/java/net/fender/pogo/IndividualValues.java
@@ -1,6 +1,5 @@
 package net.fender.pogo;
 
-import java.math.BigDecimal;
 import java.util.Objects;
 
 public class IndividualValues {
@@ -8,25 +7,25 @@ public class IndividualValues {
     public static final IndividualValues ZERO = new IndividualValues(0, 0, 0);
     public static final IndividualValues PERFECT = new IndividualValues(15, 15, 15);
 
-    private final BigDecimal attack;
-    private final BigDecimal defense;
-    private final BigDecimal stamina;
+    private final int attack;
+    private final int defense;
+    private final int stamina;
 
     public IndividualValues(int attack, int defense, int stamina) {
-        this.attack = BigDecimal.valueOf(attack);
-        this.defense = BigDecimal.valueOf(defense);
-        this.stamina = BigDecimal.valueOf(stamina);
+        this.attack = attack;
+        this.defense = defense;
+        this.stamina = stamina;
     }
 
-    public BigDecimal getAttack() {
+    public int getAttack() {
         return attack;
     }
 
-    public BigDecimal getDefense() {
+    public int getDefense() {
         return defense;
     }
 
-    public BigDecimal getStamina() {
+    public int getStamina() {
         return stamina;
     }
 
@@ -35,13 +34,13 @@ public class IndividualValues {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         IndividualValues that = (IndividualValues) o;
-        return attack.intValue() == that.attack.intValue() &&
-                defense.intValue() == that.defense.intValue() &&
-                stamina.intValue() == that.stamina.intValue();
+        return attack == that.attack &&
+                defense == that.defense &&
+                stamina == that.stamina;
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(attack.intValue(), defense.intValue(), stamina.intValue());
+        return Objects.hash(attack, defense, stamina);
     }
 }
